[core] add the simple_future/simple_promise concurrency primitives
The existing code relies on Boost.Fiber future and promise synchronization
primitives. This adds a wrapper around the low-level C synch primitives
that has a similar interface.

They are called "simple_future" and "simple_promise" to emphasize the
fact that they do not implement the same (richer) interface as the
C++11 std::future and std::promise