[core][fiber] add the initial fiber interface
This adds the documentation for fibers, the header (with no
implementation), and the code for an example program.

The rationale is to review the interface separately from the
implementation. This has been pruned from a branch with a fully
functioning implementation that can run the monad command and
the TrieDb test suite.

The demo program (fiber-hello.c) compiles but does not link,
since the implementation is missing.