[core] add a MONAD_ABORT macro, to prefer over MONAD_ASSERT(false)
Occasionally we flow to a place in the code where we must die, where there
is not obvious choice for the exact expression to assert is false. There are
currently about 20 calls to `MONAD_ASSERT(false)` when this happens, which
is not very informative.

This adds a MONAD_ABORT macro. The difference over the abort() function
in <stdlib.h> is that it must include a printf-style message that explains
why we cannot continue. It includes the same source location and stack
backtrace machinery as the assertion failure function.