Implement `async_signal_handling`, a modern asynchronous signal handling implementation to replace our direct use of `signal()` and `sigaction()`.
This should remedy the problem on Jenkins where cancelled jobs
do not necessarily cause the process to exit, which then causes
accumulation of resource consuming CI jobs to accumulate.

This should also make progress in addressing -- and possibly has
just remedied in full -- the uninterruptible sleeping process
problem whereby after a `kill -9` of a monad process refusing to
exit it enters a state of being unkillable, hanging around until
the machine reboots. If this doesn't fix that problem, I have
left open places where instrumentation and logging can be added
so we can diagnose the sequence of events which cause the issue.
However, I personally think that because this PR disables POSIX
signal handling entirely, I think it likely that the unkillable
process problem should now disappear.

The changeset is larger than it would be otherwise as I hoisted
the `make_scope_exit()` polyfill out of `libs/async/detail` into
`libs/core`.

However, the implementation is also quite large in general to
the extent that some may feel it is overengineered. The cause is
because we are replacing async signal handling **in full**. The
POSIX async signal handling mechanism is disabled, and we
reimplement it using a dedicated signal handling thread and
proprietary Linux facilities. The reimplementation means that async
signal handlers no longer have constraints on what operations
they may perform -- they now run in the signal handling thread
and so can do anything a normal thread can do. They can allocate
memory, write to logging such as Quill, write to files etc.
This I think will be a boon down the line e.g. use of `SIGUSR1`
to interact with a long running Monad service.

We also implement a watchdog facility whereby if a signal
handler does not act on a signal received within thirty seconds,
the process is hard terminated. This means even faulty signal
handling code will no longer result in processes not dying when
they've been told to.

All existing use of async signal handlers and `alarm()` have
been upgraded to use `async_signal_handling` instead.