Use pool chunks to store the DB version history.
The existing 64k DB version history on disk format is still supported.
If a new DB is created, and there are enough spare read-write blocks
in the storage pool, two of those blocks are used to store the DB
version history instead. The 64k DB version history storage in this
case is used to store the list of chunks which make up the DB version
history, thus offering the option of a future extension point if it
became necessary.

Two copies of the DB history are still stored, and the dirty flag
is still held for each during update. Each chunk contains the two
copies. Storage could be extended, if desired, to 100% of the storage
pool.