Remove compaction offsets in db_metadata
Introduce a top-level root update with an empty prefix, making the root
a leaf node whose value stores the serialized compaction offsets, which
are also offsets before which data can be safely released upon version
invalidation.

To ensure correct garbage collection for out-of-order upserts while
preserving in-order version expiration, the value is set only once
during the first upsert of a given version. Additionally, the rule for
advancing compaction offsets is updated:
- compaction offsets strictly increase for versions dependent on prior
  ones,
- compaction offsets remain 0 for versions independent of historical
  versions, meaning no chunks are released when these versions expire.

With the above design, we no longer need compaction offsets in
db_metadata, therefore removed.

This approach guarantees the correctness of chunk releases, even when
inserts occur out of order.