BlockHashChain: Add data structure to support multiple BlockHashBuffer proposal chains
The high level design is as follows:

The BlockHashBufferProposal is an unfinalized chain. It maintains a
reference to the latest BlockHashBuffer. Any proposed hashes are kept as
a vector of deltas. On finalize, the winning chain is merged into the
parent BlockHashBuffer. In practice, the deltas vector should be
relatively small, with only 1-2 proposals buffered up before consensus
signals finalization.

BlockHashChain: Add Tests

Three test cases:

1. simple_chain: mimics the replay_ethereum use case. There are no
   forks. We submit a proposal and immediately finalize.
2. fork: Demonstrates how forked chains maintain their own state.
3. keep_latest_duplicate: Tests the edge case whereby consensus may
   submit the same round twice. Always keep the latest.