BlockHashChain: Add data structure to support multiple BlockHashBuffer proposal chains
The high level design is as follows:

The BlockHashBufferProposal is an unfinalized chain. It maintains a
reference to the latest BlockHashBuffer. Any proposed hashes are kept as
a vector of deltas. On finalize, the winning chain is merged into the
parent BlockHashBuffer. In practice, the deltas vector should be
relatively small, with only 1-2 proposals buffered up before consensus
signals finalization.

Co-authored-by: Alexander Lee <alex@monad.xyz>