monad: Populate BlockHeader with execution results
Currently, monad is using the values in the `BlockHeader` struct and
passing that directly to commit. However, this doesn't actually work
outside of the replay_ethereum usecase. The blocks consensus produces
don't have real data for the following fields.

  * gas_used
  * ommers_hash
  * logs_bloom

  * state_root
  * receipts_root
  * transactions_root
  * withdrawals_root

The first three are computed by execution before commit. The last four
are computed by commit upon upserting into TrieDb.

To support this workflow, add two methods to Chain.
  1. on_pre_commit_outputs():
     * Replay: Validates gas_used, ommers_hash, logs_bloom
     * Monad: Validates gas_used doesn't exceed gas_limit and populates
       the fields.

  2. on_post_commit_outputs():
    * Replay: Validates the state roots match the header
    * Monad: Populates header with the state roots

In a future PR we could perhaps consider separate types for an
`InputBlockHeader`, which contains all the read only fields, and thi
could be passed to execute_block(). `on_post_commit_headers()` would
return an OutputBlockHeader, which is a union of the `InputBlockHeader`
and the aforementioned output fields. This would be a clearer
separation.