Don't store merkle roots in header trie.
There are two options for storing headers correctly. Either we do
two-stage commits on the header, which is non-atomic, or we store a
subset of the header in triedb. This diff opts for the latter.o

Add `encode_block_header_db` and `decode_block_header_db` for storing
the header in the DB. We store the Eth header sans the state root,
withdrawals root, transactions root, and receipts root. Those are
stored in the data section of each table's subtrie root.

Add TrieDb::read_header(). It queries the header from the DB, decodes
it, and populates the merkle roots accordingly.

Set `parent_hash` in Chain::on_pre_commit_outputs(), and verify the
parent hash (now queried from TrieDb) is correct.