chain: Update monad devnet and monad testnet to CANCUN
The MonadConsensusHeader has the following cancun fields:

    EIP-4844: blob_gas_used
    EIP-4844: excess_blob_gas
    EIP-4788: parent_beacon_block_root

To prevent execution from failing with the FieldBeforeFork error, set
both the devnet and testnet chains to EVMC_CANCUN.

Execution currently supports a subset of cancun. All EIPs are
implemented except for EIP-4844 precompiles. For these to be useful,
this requires a nontrivial lift on the consensus side. For now, enable
the execution cancun subset. This will allow users to additionally
access transient storage.