Reimplement node writers such that:
- There is no longer any possibility of recursion during node buffer
writeout, thus guaranteeing order of node buffers written equals
intended order.

- New implementation is extensively commented in a step by step
fashion making it much easier to audit for correctness and to adjust
in the future.

- As we now allocate all necessary write buffers in advance of need,
write buffer allocation pressure will increase from before.

- If a node's value needs more write buffers than there are, and
therefore the write could not be done without possibility of
recursion, we now do a synchronous blocking write implementation
only in this case. This would only occur in very rare cases.

Obviously this reimplementation is very considerably longer in terms
of source code than the previous one. The previous one did a lot
of looping back into itself using a state machine to change what
each loop does. This kind of code is complex to audit and follow,
so this reimplementation bangs out the implementation with no
state machine and only a single loop. Yes this reimplementation is
now much longer, but now its logic is far easier to read and
understand (especially with the comments).