Add restore prefix implementation
Modify upsert api to allow subtree deletion and subtree insertion.
This option is enabled by setting prefix_update to true.

Restore prefix is to be used by statesync when inconsistency in subtree
data is detected to revert to previous state and restart transfer.

Add statesync client restore prefix api. When called, restore prefix to
previous version and resubmit request for the prefix.