Limit the memory allocated for traverse clones during traverse
Keep track of number of traverse instances created per tree level. If
over the limit, stop processing nodes at previous level and add them to
delay queue. Once under the limit again, resume processing of previous
level nodes.

Since the number of levels in the tree is finite, this puts a limit on
the amount of memory a traverse can consume.