Implement in order traversals
Keep a sorted by key list of TraverseMachine instances that have incomplete
ios. Call visit function in the list order. If an io completes and
traverse is not in the front of the list, delay calling visit function.

Keep the list of pending ios sorted by key.

Implement merging of overlapping ios on the pending list to improve
traverse performance by reducing number of ios necessary.

Miscellaneous performance improvements.