[monad] carve out most of runloop_ethereum
Previously, runloop_ethereum did a lot of work. This splits it into
two functions:

  - process_ethereum_block does most of the "real work" that the big
    runloop_ethereum used to do

  - runloop_ethereum is now an "outer shell": it loads the next block
    from the BlockDb, makes the fake Monad consensus header, calls
    process_ethereum_block to do real work, and updates the log_tps
    stuff afterwards

The reason for doing this is that the event recording PR will add more
"before and after" decorations to `runloop_ethereum`, while mostly
leaving the core `process_ethereum_block` alone. It is extremely hard
to review without separating the refactor from the new functionality.