[cmake] option for the two C++ libs used from Rust to be shared
Before last week, all libraries were explicitly built as STATIC, except
monad_rpc and monad_statesync. This had a nice property: when
BUILD_SHARED_LIBS=ON was set in Rust, the `ldd` dynamic link dependency
list for monad-node only had libmonad_rpc.so and libmonad_statesync.so as
shared library dependencies.

Once STATIC was removed from most libraries, a build with
BUILD_SHARED_LIBS=ON builds _everything_ shared, and monad-node has 10+
different libraries as dependencies (libmonad_core.so, libmonad_async.so,
libmonad-vm-*.so, etc.), all of which have to be explicitly copied around
when making the docker image.

This change helps restore the old system somewhat: now we can choose all
static, all shared, or "always build the C++ libraries used from Rust
as shared." If we do the latter and BUILD_SHARED_LIBS=OFF (the default),
we'll force -fPIC on and get the old behavior of a small number of
top-level dependencies.