[mpt] Change virtual_offset_t hasher
  * Switch from fnv1 to wyhash, which ships with ankerl
  * Since virtual_chunk_offset_t is an alias for uint64, tell ankerl
    it's a primitive and is a high quality hash that doesn't require an
    extra mixing step.

I noticed Max is eliding calls to the LRU by storing it's pointer in
cache in the in memory node. This implies we are CPU bound when reading
data inside the LRU.

I ran a benchmark on an LRU with 50k entries that does 10 million reads.

With these changes, the time to perform that operations goes from 1490ms
to 980ms, or about 40% speedup.

I made this the default hasher for chunk_offset_t so other data
structures that use the chunk_offset_t_hasher, such as the inflights
map, can benefit from these improvements.