Merge clean_storage and post transition cleanup
When working on the shrinker, I found that the
post transition cleanup removes empty accounts
that may be referred by the journal. As a result,
when calling rollback() after the transition, the
journal may try to lookup the account that was
deleted by the post transition cleanup and throw
an exception. The solution to this problem is to
move the cleanup to after the rollbacks.