[call-tracer] CallFrame::Log stores receipt_index
There are two changes here:

- State::store_log returns the index in the receipt's log array that
  the log will have, and it becomes [[nodiscard]] because whomever is
  calling it shouldn't forget to propagate it to the call tracer

- CallTracer::on_log is passed the receipt index, and records it into
  the CallFrame

The rationale is to support the call frame<->log association in the
execution events.

It does not make sense to record the logs twice, especially if they
are large. They will instead be recorded by their receipt index,
which is the same as the event recording order.

The receipt_index is not added to the RLP-encoded form of the
CallFrameLog, since it stores the entire log anyway. This de-duplication
is only for the efficiency of the event recording, and we don't want the
database or the RPC-side decoder to care about it. Consequently, the
`receipt_index` field is not round-trippable through RLP, has type
std::optional<size_t>, and decode-constructs with its value equal to
std::nullopt.